Рефакторит form-photo-upload
